Reverting graphql util
