+++ graphql/task_resolver.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	gqlgen "github.com/99designs/gqlgen/graphql"
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/apimodels"
 	"github.com/evergreen-ci/evergreen/graphql/loaders"
@@ -876,31 +875,6 @@ func (r *taskResolver) Tests(ctx context.Context, obj *restModel.APITask, opts *
 		TotalTestCount:    taskResults.Stats.TotalCount,
 		FilteredTestCount: utility.FromIntPtr(taskResults.Stats.FilteredCount),
 	}, nil
-}
-
-func shouldDecorateTestQuarantineStatus(ctx context.Context) bool {
-	if !gqlgen.HasOperationContext(ctx) {
-		return true
-	}
-	return collectedFieldPathContains(gqlgen.GetOperationContext(ctx), gqlgen.CollectFieldsCtx(ctx, nil), []string{"testResults", "isManuallyQuarantined"})
-}
-
-func collectedFieldPathContains(opCtx *gqlgen.OperationContext, fields []gqlgen.CollectedField, path []string) bool {
-	if len(path) == 0 {
-		return true
-	}
-	for _, field := range fields {
-		if field.Name != path[0] {
-			continue
-		}
-		if len(path) == 1 {
-			return true
-		}
-		if collectedFieldPathContains(opCtx, gqlgen.CollectFields(opCtx, field.Selections, nil), path[1:]) {
-			return true
-		}
-	}
-	return false
 }
 
 // TotalTestCount is the resolver for the totalTestCount field.
